+++ fresh.gen.ts
@@ -5,12 +5,14 @@
 import * as $0 from './routes/_app.tsx';
 import * as $1 from './routes/index.tsx';
 import * as $2 from './routes/md.tsx';
+import * as $3 from './routes/test/input.tsx';
 
 const manifest = {
   routes: {
     './routes/_app.tsx': $0,
     './routes/index.tsx': $1,
     './routes/md.tsx': $2,
+    './routes/test/input.tsx': $3,
   },
   islands: {},
   baseUrl: import.meta.url,